Commit
The project is named so as it helps people to store their important notes safe and help you remember you all the important points. The notes made gets stored in the local memory and remains on the site until you delete it. It is made with the help of html, css and js.